get tags and node IDs working
The tags are not always on the first span message, but seem to always be the
same for all of the messages.  So just populate the span with the first
non-empty tag.

And then we can extract the node id from the tag.  The code is ugly right now
but we have a time cruch.